requirements.txt
discord.py
disputils
python-decouple
pyrebase4
pytz